requirements.txt
pydantic
python-dotenv
numpy
torch
tqdm
ffmpeg
ffmpeg-python
transformers
datasets
evaluate
langchain
langchainhub
langchain-openai
llama-cpp-python
google-search-results
playwright
docarray
faiss-gpu
arxiv
pymupdf
optimum
auto-gptq
autoawq
bitsandbytes
peft